[types] add `CommunityReputation`
